Added support for non-rectangular desktop configurations
The coordinate system for the desktop appears to be the encolosing
rectangle for all screens. In this case, the minimum y for a given
monitor may not be 0. Similarly, for dual monitor setups, the
minimum x for a monitor is nonzero.

This seems to also fix the issues with top-bar height and most
likely alternative configurations for dock location.